refactor(activity): statusLabel lives with the other label helpers
A component file may only export components, so the status words move to
jobLabels beside the type, transport and skip-reason ones.
