+++ frontend/src/components/StatusBadge.tsx
@@ -1,36 +1,8 @@
 import React from 'react'
 import { Chip, Tooltip } from '@mui/material'
 import { useTranslation } from 'react-i18next'
-import type { TFunction } from 'i18next'
 import { CircleMinus } from 'lucide-react'
-
-/** The words for a status, wherever it is shown: a badge, a tooltip, or the
- *  accessible name of a status dot that shows no words of its own. */
-export function statusLabel(status: string, t: TFunction): string {
-  switch (status.toLowerCase()) {
-    case 'completed':
-      return t('status.completed')
-    case 'completed_with_warnings':
-      return t('status.completedWithWarnings')
-    case 'needs_backup':
-      return t('status.needsBackup')
-    case 'failed':
-      return t('status.failed')
-    case 'running':
-    case 'in_progress':
-      return t('status.running')
-    case 'pending':
-      return t('status.pending')
-    case 'queued':
-      return t('status.queued')
-    case 'cancelled':
-      return t('status.cancelled')
-    case 'skipped':
-      return t('status.skipped')
-    default:
-      return status.charAt(0).toUpperCase() + status.slice(1)
-  }
-}
+import { statusLabel } from './jobs/jobLabels'
 
 interface StatusBadgeProps {
   status: string
